rename into ci.py and python ci.py deploy
