+++ .dockerignore
@@ -11,4 +11,4 @@ TODO
 deploy.sh
 logs/*
 .aws
-deploy.py
+ci.py
