+++ .gitignore
@@ -75,4 +75,4 @@ opentf_deploy/image_tag.txt
 opentf_deploy/.terraform/providers/registry.opentofu.org/hashicorp/local
 logs/*
 .aws
-deploy.py
+ci.py
